refactor(system-settings): 将 ProductImprovementDialog 归入系统设置模块
ProductImprovementDialog 仅在系统设置页面使用，将其从全局组件目录移入
系统设置模块内，同时调整其在页面中的渲染位置（移出 PageWrapper），
避免布局嵌套问题
