+++ src/renderer/pages/settings/system/components/ProductImprovementDialog.tsx
@@ -14,7 +14,7 @@ import React, { useCallback } from 'react';
 import { Button } from '@arco-design/web-react';
 import { useTranslation } from 'react-i18next';
 import { Shield, Close } from '@icon-park/react';
-import AionModal from './base/AionModal';
+import AionModal from '../../../../components/base/AionModal';
 
 interface ProductImprovementDialogProps {
   /** 是否显示弹窗 */
